requirements.txt
Bootstrap-Flask>=1.5
click
Flask
Flask-Babel
Flask-Login
Flask-Mail
Flask-Migrate
Flask-Moment
Flask-SQLAlchemy
Flask-WTF
requests
itsdangerous==1.1.0
MarkupSafe==1.1.1
python-dotenv
PyJWT
WTForms